No need to `reference` and `--composite false`
...when we don't use tsc for compilation

The references option in tsconfig.json is used for managing project
references in TypeScript's incremental build system. Since esbuild does
not perform type-checking or build orchestration, this option is not
directly utilized by Vite or esbuild.

Consequently we don't need to specify the `composite` option when we
don't use the feature.